Move login redirect to separate component
